int instead of long

+++ src/main/java/com/github/zeroeighteightzero/matn/Glyph.java
@@ -9,13 +9,13 @@ package com.github.zeroeighteightzero.matn;
  */
 public class Glyph {
 
-    public final long glyphID;
+    public final int glyphID;
     public final int size;
     public final int page;
     public final int x, y, width, height;
     public final int top, left;
 
-    Glyph(GlyphAtlas atlas, long glyphID, int size, int page, int x, int y, int width, int height, int top, int left) {
+    Glyph(GlyphAtlas atlas, int glyphID, int size, int page, int x, int y, int width, int height, int top, int left) {
         this.glyphID = glyphID;
         this.size = size;
         this.page = page;
